+++ releasenotes.md
@@ -2,3 +2,4 @@ Whats New:
 ----------------------
 Fix Core upgrade issue when running in a non docker enviroment
 Calculation tools moved under top right menu
+Password reset process added